replace state management with log_provider but haven't test yet

+++ lib/isar/model/notelog.dart
@@ -10,7 +10,7 @@ class NoteLog {
 
   String? labelMood;
 
-  int? numericMood;
+  double? numericMood;
 
   late DateTime date;
